refactor execute
